pyproject.toml
[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.30.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "langchain>=0.3.23",
    "langchain-anthropic>=0.3.12",
    "langchain-core>=0.3.54",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.1.2",
    "langchain-openai>=0.3.14",
    "numpy>=2.2.4",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "transformers>=4.51.3",
]
